[fix] safe-npm spinner issue
